# docs/swagger.yaml
basePath: /api
definitions:
  models.Comment:
    properties:
      author_id:
        type: integer
      author_name:
        type: string
      can_delete:
        type: boolean
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      post_id:
        type: integer
    type: object
  models.Post:
    properties:
      author_id:
        type: integer
      author_name:
        type: string
      can_edit:
        type: boolean
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      thread_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: API для форума
  title: Forum Service API
  version: "1.0"
paths:
...
  /posts:
    get:
      description: Отображает HTML-страницу со списком всех постов.
      produces:
      - text/html
      responses:
        "200":
          description: HTML страница
          schema:
            type: string
        "500":
          description: ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Список постов
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Создаёт новый пост в указанном треде. Доступно только авторизованным
        пользователям.
      parameters:
      - description: Данные для создания поста
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: пользователь не аутентифицирован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: нет прав для создания поста в этом треде
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать новый пост
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Удаляет пост. Доступно только автору поста или администратору.
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: invalid post ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: no permission to delete this post
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить пост
      tags:
      - posts
    get:
      description: Отображает HTML-страницу с информацией о посте.
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML страница
          schema:
            type: string
        "400":
          description: Неверный ID поста
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пост не найден
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Показать пост
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Обновляет информацию о посте. Доступно только автору поста или
        администратору.
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления поста
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: invalid post ID или неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: no permission to update this post
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить пост
      tags:
      - posts
  /posts/{id}/comments:
    get:
      description: Возвращает информацию о посте и все комментарии к нему.
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'post: информация о посте, comments: список комментариев'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid post ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: post not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить пост с комментариями
      tags:
      - posts
  /posts/{id}/edit:
    get:
      description: Отображает HTML-страницу с формой для редактирования поста.
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML страница
          schema:
            type: string
        "400":
          description: Неверный ID поста
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пост не найден
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Показать форму редактирования поста
      tags:
      - posts
...
schemes:
- http
- https
swagger: "2.0"